CNV-71659: Address review nits on disk row actions
Use getNamespace and getDataVolumeName, extract shared DiskRowActions
types, reuse the CDROM dropdown item, and guard getVolumes before find.
