+++ src/views/virtualmachines/details/tabs/configuration/storage/components/tables/disk/types.ts
@@ -1,0 +1,13 @@
+import {
+  type V1VirtualMachine,
+  type V1VirtualMachineInstance,
+} from '@kubevirt-ui-ext/kubevirt-api/kubevirt';
+import { type DiskRowDataLayout } from '@kubevirt-utils/resources/vm/utils/disk/constants';
+
+export type DiskRowActionsProps = {
+  customize?: boolean;
+  obj: DiskRowDataLayout;
+  onDiskUpdate?: (updatedVM: V1VirtualMachine) => Promise<V1VirtualMachine>;
+  vm: V1VirtualMachine;
+  vmi?: V1VirtualMachineInstance;
+};
